added example for tsdb
